collateMyCloudData-master/requirements.txt
selenium==3.141.0
pandas==1.2.4
protobuf==3.17.1
python_dateutil==2.8.1
google-cloud-storage
openpyxl
